requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.1
frozenlist==1.4.1
future==0.18.3
gevent==23.9.1
greenlet==3.0.2
idna==3.6
inflection==0.5.1
multidict==6.0.4
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.4
tldextract==5.1.1
twilio==8.11.0
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
